Relocate getQueryModelExportParams
